+++ Frends.AzureBlobStorage.ListBlobsInContainer/CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Updated
 - Azure.Identity from 1.8.1 to 1.10.2
 - Azure.Storage.Blobs from 12.11.0 to 12.13.0
+### Fixed
+- Documentation fixes.
+- Changed the Task to throw ArgumentException instead of generic Exception.
 
 ## [1.1.1] - 2023-12-11
 ### Added